全面整理代码，去除编译警告

+++ api/include/mk_events.h
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLMediaKit project authors. All Rights Reserved.
  *
- * This file is part of ZLMediaKit(https://github.com/xiongziliang/ZLMediaKit).
+ * This file is part of ZLMediaKit(https://github.com/xia-chu/ZLMediaKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -148,8 +148,8 @@ typedef struct {
      * @param is_player 客户端是否为播放器
      */
     void (API_CALL *on_mk_flow_report)(const mk_media_info url_info,
-                                       uint64_t total_bytes,
-                                       uint64_t total_seconds,
+                                       size_t total_bytes,
+                                       size_t total_seconds,
                                        int is_player,
                                        const mk_sock_info sender);
 } mk_events;
