+++ src/Extension/H265.cpp
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLMediaKit project authors. All Rights Reserved.
  *
- * This file is part of ZLMediaKit(https://github.com/xiongziliang/ZLMediaKit).
+ * This file is part of ZLMediaKit(https://github.com/xia-chu/ZLMediaKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -14,7 +14,7 @@
 
 namespace mediakit{
 
-bool getHEVCInfo(const char * vps, int vps_len,const char * sps,int sps_len,int &iVideoWidth, int &iVideoHeight, float  &iVideoFps){
+bool getHEVCInfo(const char * vps, size_t vps_len,const char * sps,size_t sps_len,int &iVideoWidth, int &iVideoHeight, float  &iVideoFps){
     T_GetBitContext tGetBitBuf;
     T_HEVCSPS tH265SpsInfo;	
     T_HEVCVPS tH265VpsInfo;
@@ -22,7 +22,7 @@ bool getHEVCInfo(const char * vps, int vps_len,const char * sps,int sps_len,int 
         memset(&tGetBitBuf,0,sizeof(tGetBitBuf));	
         memset(&tH265VpsInfo,0,sizeof(tH265VpsInfo));
         tGetBitBuf.pu8Buf = (uint8_t*)vps+2;
-        tGetBitBuf.iBufSize = vps_len-2;
+        tGetBitBuf.iBufSize = (int)(vps_len-2);
         if(0 != h265DecVideoParameterSet((void *) &tGetBitBuf, &tH265VpsInfo)){
             return false;
         }
@@ -32,7 +32,7 @@ bool getHEVCInfo(const char * vps, int vps_len,const char * sps,int sps_len,int 
         memset(&tGetBitBuf,0,sizeof(tGetBitBuf));
         memset(&tH265SpsInfo,0,sizeof(tH265SpsInfo));
         tGetBitBuf.pu8Buf = (uint8_t*)sps+2;
-        tGetBitBuf.iBufSize = sps_len-2;
+        tGetBitBuf.iBufSize = (int)(sps_len-2);
         if(0 != h265DecSeqParameterSet((void *) &tGetBitBuf, &tH265SpsInfo)){
             return false;
         }
@@ -47,7 +47,7 @@ bool getHEVCInfo(const char * vps, int vps_len,const char * sps,int sps_len,int 
 }
 
 bool getHEVCInfo(const string &strVps, const string &strSps, int &iVideoWidth, int &iVideoHeight, float &iVideoFps) {
-    return getHEVCInfo(strVps.data(),strVps.size(),strSps.data(),strSps.size(),iVideoWidth,iVideoHeight,iVideoFps);
+    return getHEVCInfo(strVps.data(), strVps.size(), strSps.data(), strSps.size(), iVideoWidth, iVideoHeight,iVideoFps);
 }
 
 Sdp::Ptr H265Track::getSdp() {
