+++ src/Http/HlsParser.h
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2020 The ZLMediaKit project authors. All Rights Reserved.
  *
- * This file is part of ZLMediaKit(https://github.com/xiongziliang/ZLMediaKit).
+ * This file is part of ZLMediaKit(https://github.com/xia-chu/ZLMediaKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -68,7 +68,7 @@ public:
     /**
      * #EXT-X-MEDIA-SEQUENCE字段值，该m3u8序号
      */
-    int getSequence() const;
+    int64_t getSequence() const;
 
     /**
      * 内部是否含有子m3u8
